requirements.txt
# Python dependencies for Field Prime Viz backend
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
numpy==2.1.3
scipy==1.15.2
pillow==11.3.0
torch==2.6.0
pandas==2.3.2
scikit-learn==1.6.1
python-socketio==5.11.1